Inline the CompilerCaching extension; delegate inference to it on 1.11+.
CompilerCaching becomes a strong dep (it loads as an empty shell on 1.10,
so there's no overhead). The `GPUCompilerCompilerCachingExt` extension —
which existed just to wire the parametric `CC.finish!` override that
attaches a `CachedResult{V}` to every inferred CodeInstance — moves into
`jlgen.jl` directly, gated on `HAS_INTEGRATED_CACHE`. Same code, one less
file.

With CompilerCaching available unconditionally we can also drop the inline
copies of its inference machinery from `jlgen.jl`:

- `drive_inference!` on 1.11+ is now a two-line delegation to
  `CompilerCaching.typeinf!` + `get(cache, mi, nothing)`. The 1.10
  implementation (which talks to the per-interpreter `CodeCache`) moves to
  `deprecated.jl`.
- `collect_codeinfos` / `_ci_codeinfo` go away; the single call site in
  `compile_method_instance` calls `CompilerCaching.get_codeinfos` directly.
- `StackedMethodTable` is re-exported from CompilerCaching on 1.11+; the
  1.10 variant (with the older `MethodMatchResult`-shape `findall`) moves
  to `deprecated.jl`.

Net result: ~200 lines deleted from `jlgen.jl`, no behavior change. All
1.10-only code is now in `deprecated.jl`, ready to disappear in one diff
when 1.10 support drops.
