+++ Project.toml
@@ -7,6 +7,7 @@ authors = ["Tim Besard <tim.besard@gmail.com>"]
 projects = ["test"]
 
 [deps]
+CompilerCaching = "9db33cc3-5358-4881-8759-fa4194144afd"
 ExprTools = "e2ba6199-217a-4e67-a87a-7c52f15ade04"
 InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
 LLVM = "929cbde3-209d-540e-8aea-75f648917ca0"
@@ -21,12 +22,8 @@ UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 
 [weakdeps]
 AMDGPU_LLVM_Backend_jll = "cc5c0156-bd05-5a77-8a68-bb0aafb29019"
-CompilerCaching = "9db33cc3-5358-4881-8759-fa4194144afd"
 LLVMDowngrader_jll = "f52de702-fb25-5922-94ba-81dd59b07444"
 NVPTX_LLVM_Backend_jll = "ef6e0fe3-e6ef-59c0-bde6-4989574699e0"
-
-[extensions]
-GPUCompilerCompilerCachingExt = "CompilerCaching"
 
 
 [compat]
